Move controllers out of presroc folder
