+++ app/controllers/index.js
@@ -8,12 +8,12 @@ const AuthorisedSystemsController = require('./admin/authorised_systems.controll
 const CustomersController = require('./admin/customers.controller')
 const DatabaseController = require('./admin/health/database.controller')
 const NotSupportedController = require('./not_supported.controller')
-const PresrocBillRunsController = require('./presroc/bill_runs.controller')
-const PresrocBillRunsInvoicesController = require('./presroc/bill_runs_invoices.controller')
-const PresrocBillRunsLicencesController = require('./presroc/bill_runs_licences.controller')
-const PresrocBillRunsTransactionsController = require('./presroc/bill_runs_transactions.controller')
-const PresrocCalculateChargeController = require('./presroc/calculate_charge.controller')
-const PresrocCustomerDetailsController = require('./presroc/customer_details.controller')
+const PresrocBillRunsController = require('./bill_runs.controller')
+const PresrocBillRunsInvoicesController = require('./bill_runs_invoices.controller')
+const PresrocBillRunsLicencesController = require('./bill_runs_licences.controller')
+const PresrocBillRunsTransactionsController = require('./bill_runs_transactions.controller')
+const PresrocCalculateChargeController = require('./calculate_charge.controller')
+const PresrocCustomerDetailsController = require('./customer_details.controller')
 const RegimesController = require('./admin/regimes.controller')
 const TestBillRunsController = require('./admin/test/test_bill_runs.controller')
 const TestCustomerFilesController = require('./admin/test/test_customer_files.controller')
